Initiate the git repository
This is the initial commit for the universal compiler project. This
commit can be used as a reference for how commits should be formatted.
The commit message rules are:

 - Separate the title from the body with a blank line
 - Limit the title to 72 characters
 - Capitilize the title
 - Do not end the title with a period
 - Use the imperative mood in the title
 - The maximum line length is 72 characters
 - Always state what was changed and why.

You can test that the title is appropriately worded if it can
appropriately be substituted into the text "This commit will ???".

This commit includes an empty `.gitignore` so that it will not be
forgotten later. This should prevent any unusual objects in the git
database.

This commit also initializes an empty `README.md` containing a basic
TODO comment. This is also so that it is not forgotten as this project
is a tool and must be thoroughly documented to be effective.