Add make system
This commit prepares the compilation toolchain for smooth integration.

The gitignore was updated with all files that currently may need to be
ignored.

The build directory was created with a makefile and symbolic links to
some source directories.

The src directory was given its makefile and a symbolic link to the
build makefile in order to cleanly build the ada code without dumping
unusual files in the working src directory.

A simple "Hello, World!" style test program was written to ensure that
the build system is working.